描画処理と当たり判定の処理の追加

+++ plant_parameters.hpp
@@ -214,7 +214,7 @@ public:
 		double ligand = Pow(getMood(), getPlantLevel()); // リガンド濃度
 		double dissociation = Pow(20, getPlantLevel()); // 解離定数
 		
-		mood_magnification = Ceil(ligand / (dissociation + ligand) * 10.0);
+		mood_magnification = int(Ceil(ligand / (dissociation + ligand) * 10.0));
 		return 0;
 	}
